Enhance scalar functions with span-based optimizations and memory management
- Added `NeedsQuoting`, `QuoteIfNeeded`, and `IsBareIdentifier` methods to `SqlIdentifier` for efficient identifier handling.
- Refactored `ArrayJoinFunction`, `FormatFunction`, `QuoteIdentFunction`, `QuoteLiteralFunction`, `QuoteNullableFunction`, and others to utilize `Span<char>` for reduced allocations and improved performance.
- Introduced `Execute` methods in various functions to handle `ReadOnlySpan<DataValue>` arguments, ensuring better memory management with rented buffers.
- Updated tests in `JsonFunctionTests` and added new tests in `ScalarFunctionStoreAwareBatch2Tests` to validate changes and ensure correctness.

+++ tests/DatumIngest.Tests/Functions/JsonFunctionTests.cs
@@ -252,7 +252,7 @@ public class JsonFunctionTests
         DataValue result = function.Execute([
             DataValue.FromJsonValue("[1, 2, 3, 4, 5]")
         ]);
-        Assert.Equal(5f, result.AsFloat32());
+        Assert.Equal(5, result.AsInt32());
     }
 
     [Fact]
@@ -263,7 +263,7 @@ public class JsonFunctionTests
             DataValue.FromJsonValue("{\"items\": [\"a\", \"b\"]}"),
             DataValue.FromString("items")
         ]);
-        Assert.Equal(2f, result.AsFloat32());
+        Assert.Equal(2, result.AsInt32());
     }
 
     [Fact]
@@ -283,7 +283,7 @@ public class JsonFunctionTests
         DataValue result = function.Execute([
             DataValue.FromJsonValue("[]")
         ]);
-        Assert.Equal(0f, result.AsFloat32());
+        Assert.Equal(0, result.AsInt32());
     }
 
     [Fact]
